finished

+++ TodoList/Views/ListView.swift
@@ -11,20 +11,27 @@ struct ListView: View {
     @EnvironmentObject var listViewModel : ListViewModel
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items){item in
-               ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear(duration: 0.2)){
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack{
+            if listViewModel.items.isEmpty{
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeInOut))
+            }else{
+                List{
+                    ForEach(listViewModel.items){item in
+                       ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear(duration: 0.2)){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(.plain)
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(.plain)
-        .navigationTitle("Todo List 📝")
+        .navigationTitle("Todo List")
         .navigationBarItems(
             leading: EditButton(),
             trailing: NavigationLink(
